refactor appDelegate extension into more specific ones
